Conflicts:
	bindings/json/pom.xml
	bindings/kvp/pom.xml
	bindings/pom.xml
	bindings/pox/pom.xml
	bindings/rest/code/pom.xml
	bindings/rest/pom.xml
	bindings/rest/xml/pom.xml
	bindings/soap/pom.xml
	coding/json-common/pom.xml
	coding/json/pom.xml
	coding/ows-v110/pom.xml
	coding/pom.xml
	coding/sensorML-v101/pom.xml
	coding/sos-v100/pom.xml
	coding/sos-v20/pom.xml
	coding/wml-v20/pom.xml
	core/admin/pom.xml
	core/api/pom.xml
	core/cache/pom.xml
	core/gc-dao/pom.xml
	core/pom.xml
	core/profile/coding/pom.xml
	core/profile/pom.xml
	core/profile/xml/pom.xml
	core/sqlite-config/pom.xml
	core/test/pom.xml
	data/pom.xml
	extensions/do/core/pom.xml
	extensions/do/hibernate/pom.xml
	extensions/do/json/pom.xml
	extensions/do/kvp/pom.xml
	extensions/do/pom.xml
	extensions/do/xml/pom.xml
	extensions/gda/hibernate/pom.xml
	extensions/gda/operation/pom.xml
	extensions/gda/pom.xml
	extensions/pom.xml
	hibernate/common/pom.xml
	hibernate/dao/pom.xml
	hibernate/datasource/common/pom.xml
	hibernate/datasource/h2/pom.xml
	hibernate/datasource/mysql/pom.xml
	hibernate/datasource/oracle/pom.xml
	hibernate/datasource/pom.xml
	hibernate/datasource/postgres/pom.xml
	hibernate/feature/pom.xml
	hibernate/h2/pom.xml
	hibernate/mappings/pom.xml
	hibernate/pom.xml
	hibernate/session-factory/pom.xml
	operations/core-v100/pom.xml
	operations/core-v20/pom.xml
	operations/enhanced-v20/pom.xml
	operations/pom.xml
	operations/resultHandling-v20/pom.xml
	operations/transactional-v20/pom.xml
	pom.xml
	spring/admin-controller/pom.xml
	spring/client-controller/pom.xml
	spring/common-controller/pom.xml
	spring/common-controller/src/main/java/org/n52/sos/web/ControllerConstants.java
	spring/install-controller/pom.xml
	spring/pom.xml
	spring/views/pom.xml
	spring/views/src/main/webapp/WEB-INF/views/common/footer.jsp
	spring/views/src/main/webapp/WEB-INF/views/common/header.jsp
	spring/views/src/main/webapp/WEB-INF/views/index.jsp
	third-party/pom.xml
	third-party/third-party-ncname-resolver/pom.xml
	webapp/pom.xml
	webapp/src/main/webapp/static/conf/client-config.json

+++ core/api/src/main/java/org/n52/sos/ogc/om/values/QuantityValue.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2012-2014 52°North Initiative for Geospatial Open Source
+ * Copyright (C) 2012-2015 52°North Initiative for Geospatial Open Source
  * Software GmbH
  *
  * This program is free software; you can redistribute it and/or modify it
@@ -28,8 +28,6 @@
  */
 package org.n52.sos.ogc.om.values;
 
-import java.math.BigDecimal;
-
 import org.n52.sos.util.StringHelper;
 
 /**
@@ -38,7 +36,7 @@ import org.n52.sos.util.StringHelper;
  * @since 4.0.0
  * 
  */
-public class QuantityValue implements Value<BigDecimal> {
+public class QuantityValue implements Value<Double> {
     /**
      * serial number
      */
@@ -47,7 +45,7 @@ public class QuantityValue implements Value<BigDecimal> {
     /**
      * Measurement value
      */
-    private BigDecimal value;
+    private Double value;
 
     /**
      * Unit of measure
@@ -60,7 +58,7 @@ public class QuantityValue implements Value<BigDecimal> {
      * @param value
      *            Measurement value
      */
-    public QuantityValue(BigDecimal value) {
+    public QuantityValue(Double value) {
         this(value, null);
     }
 
@@ -72,18 +70,18 @@ public class QuantityValue implements Value<BigDecimal> {
      * @param unit
      *            Unit of measure
      */
-    public QuantityValue(BigDecimal value, String unit) {
+    public QuantityValue(Double value, String unit) {
         this.value = value;
         this.unit = unit;
     }
 
     @Override
-    public void setValue(BigDecimal value) {
+    public void setValue(Double value) {
         this.value = value;
     }
 
     @Override
-    public BigDecimal getValue() {
+    public Double getValue() {
         return value;
     }
