+++ hibernate/common/src/main/java/org/n52/sos/ds/hibernate/entities/interfaces/NumericObservation.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2012-2014 52°North Initiative for Geospatial Open Source
+ * Copyright (C) 2012-2015 52°North Initiative for Geospatial Open Source
  * Software GmbH
  *
  * This program is free software; you can redistribute it and/or modify it
@@ -28,8 +28,6 @@
  */
 package org.n52.sos.ds.hibernate.entities.interfaces;
 
-import java.math.BigDecimal;
-
 import org.n52.sos.ds.hibernate.entities.HibernateRelations.HasValue;
 
 /**
@@ -38,6 +36,6 @@ import org.n52.sos.ds.hibernate.entities.HibernateRelations.HasValue;
  * @since 4.0.0
  * 
  */
-public interface NumericObservation extends HasValue<BigDecimal> {
+public interface NumericObservation extends HasValue<Double> {
 
 }
